added force flag to delete fip

+++ lithops/standalone/backends/ibm_vpc/ibm_vpc.py
@@ -330,13 +330,14 @@ class IBMVPCBackend:
                     if ins_to_dlete not in deleted_instances:
                         instances_to_delete.add(ins_to_dlete)
                     if ins['name'].startswith(LITHOPS_MASTER):
-                        # clean all been triggered, delete also master floating IP
-                        interface_id = ins['network_interfaces'][0]['id']
-                        fips = self.ibm_vpc_client.list_instance_network_interface_floating_ips(
-                            ins['id'], interface_id).get_result()['floating_ips']
-                        if fips:
-                            fip = fips[0]['id']
-                            self.ibm_vpc_client.delete_floating_ip(fip)
+                        if self.config.get('force'):
+                            # forced clean all been triggered, delete also master floating IP
+                            interface_id = ins['network_interfaces'][0]['id']
+                            fips = self.ibm_vpc_client.list_instance_network_interface_floating_ips(
+                                ins['id'], interface_id).get_result()['floating_ips']
+                            if fips:
+                                fip = fips[0]['id']
+                                self.ibm_vpc_client.delete_floating_ip(fip)
 
             if instances_to_delete:
                 with ThreadPoolExecutor(len(instances_to_delete)) as executor:
